requirements.txt
behave==1.2.6
PyHamcrest==1.9.0
requests==2.22.0
selenium==3.141.0
junit2html==22
allure-behave==2.7.0


